Update OnboardingTableViewCell2 to OnboardingTableViewCell
